requirements.txt
fastapi
uvicorn
httpx
python-dotenv
beautifulsoup4
apscheduler